Add main package to web module, add annotation to Parser
